+++ Docs/CHANGELOG.md
@@ -20,10 +20,14 @@
     - Added initial speed support for pedestrians for OpenSCENARIO
     - Support for EnvironmentActions within Story (before only within Init). This allows changing weather conditions during scenario execution
     - Extended FollowLeadingVehicle example to illustrate weather changes
+    - Added support for RelativeSpeedCondition
+    - Added support for AccelerateCondition
 * Atomics:
     - WeatherBehavior to simulate weather over time
     - UpdateWeather to update weather to a new setting, e.g. sun to rain
     - UpdateRoadFriction to update the road friction while running
+    - new RelativeVelocityToOtherActor trigger condition, used to compare velocities of two actors
+    - new TriggerAcceleration trigger condition which compares a reference acceleration with the actor's one.
 ### :bug: Bug Fixes
 * Fixed initial speed of vehicles using OpenSCENARIO
 * Fixed bug causing an exception when calling BasicScenario's *_initialize_actors* with no other_actors.